DocumentPageをConsumerStatefulWidgetに変更
